requirements.txt
cycler==0.10.0
dataclasses==0.6
kiwisolver==1.2.0
matplotlib==3.2.1
numpy==1.18.3
opencv-python==4.2.0.34
Pillow==7.1.2
pyparsing==2.4.7
python-dateutil==2.8.1
scikit-image==0.17.1
scipy==1.4.1
six==1.14.0
